refactor: streamline imports and enhance cookie handling in user books API
