+++ src/hooks/api/useUserBooks.ts
@@ -1,6 +1,5 @@
+import { fetchUserBooks } from '@api/books/userBooks.service'
 import { useQuery } from '@tanstack/react-query'
-
-import { fetchUserBooks } from '@src/api/books/userBooks.service'
 
 export const useUserBooks = () => {
   return useQuery({ queryKey: ['userBooks'], queryFn: fetchUserBooks })
