+++ tests/hooks/api/useUserBooks.test.tsx
@@ -1,8 +1,7 @@
+import { useUserBooks } from '@hooks/api/useUserBooks'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import { renderHook, waitFor } from '@testing-library/react'
 import { describe, expect, test } from 'vitest'
-
-import { useUserBooks } from '@src/hooks/api/useUserBooks'
 
 describe('useUserBooks', () => {
   test('fetches user books', async () => {
